fix: accept decimals in every numeric field, and show client names in full
Decimals were blocked at three separate layers and, separately, silently
lost for anyone using a French keyboard.

Keyboards: measurements (both the editor and template-driven fields) used
`numeric`, and invoice line quantity used `number-pad` — which has no
decimal key at all. A tailor could not enter 38.5 cm. Now `decimal-pad`.

The comma. SeamFlow is used in French-speaking Cameroon, where the decimal
key types a comma. `Number("2,5")` is NaN, so fabric yardage silently
became null on the order screen and in new-order, and fabric yardage/cost
— which are numeric-as-string — sent "2,5" straight at a Postgres numeric
column. Some call sites handled this, most didn't. lib/numeric.ts is now
the one parser: comma-aware, strips NBSP/thin-space thousands separators,
and returns null rather than 0 for empty so callers can tell them apart.

Invoice quantity was integer-locked in the schema, rounded in the subtotal,
rounded again on save, and rounded a third time in the PDF. Fabric is sold
by the metre and 2.5 m is ordinary, so both the app and the API worked
around it by folding yardage into unitPrice — the bill read "1 × 7,500",
hiding both the rate and the length. Quantity is now a positive number
(line items are jsonb, so no migration) and the fabric line says
"2.5 × 3,000", which is what a client can actually check.

Client names. The order-title suggestion took only the FIRST WORD of the
client's name, so "Mme Bambot" became "Caftan — Mme" — identifying nobody,
since the first token is a courtesy title. It now uses the full name. The
PDF and the web invoice both give the name its own larger style that wraps
instead of being clipped.

Garment count stays a whole number on purpose — you don't sew 2.5 shirts.

Verified: 16 parser assertions (comma, NBSP, empty-vs-zero, negatives);
a rendered invoice showing "Mme Bambot Antoinette-Marie" in full and a
2.5 × 3,000 fabric line totalling 7,500; typecheck on app, API and web;
i18n guard 957 keys in sync.

+++ packages/schemas/src/invoice.ts
@@ -19,7 +19,10 @@ export const InvoiceLineItemSchema = z.object({
   id: z.string().min(1),
   category: InvoiceLineCategorySchema,
   description: z.string(),
-  quantity: z.number().int().positive(),
+  // Not an integer: a fabric line is priced per metre and 2.5 m is ordinary.
+  // Forcing whole numbers here is what pushed the API into folding yardage
+  // into unitPrice (see invoices.service.ts buildFabricLine).
+  quantity: z.number().positive(),
   unitPrice: z.number().nonnegative(),
 });
 export type InvoiceLineItem = z.infer<typeof InvoiceLineItemSchema>;
